+++ python/qibuild/test/test_toc.py
@@ -7,6 +7,7 @@
 """
 
 import os
+import platform
 import tempfile
 import unittest
 
@@ -141,6 +142,16 @@ class TocTestCase(unittest.TestCase):
 
     def tearDown(self):
         qisys.sh.rm(self.tmp)
+
+def test_is_build_folder_name():
+    parts = ["sys", platform.system().lower()]
+    assert qibuild.toc.is_build_folder_name(parts) == False
+    parts.append(platform.machine().lower())
+    assert qibuild.toc.is_build_folder_name(parts) == True
+    parts.append("release")
+    assert qibuild.toc.is_build_folder_name(parts) == True
+    parts.append("foo")
+    assert qibuild.toc.is_build_folder_name(parts) == False
 
 if __name__ == "__main__":
     unittest.main()